Attempting to debug issues 39 and 49 in build.jl through Travis
This commit adds a BinDeps.debug statement in an attempt to debug the joinpath issues present in the existing Travis testing seen with prior commits.  These additions will be reverted once the cause of the current build error is determined.  In an attempt to reproduce the joinpath error locally, Tony was not able to reproduce the behavior on Ubuntu, so this commit is an attempt to debug within Travis itself.