Update Ptr/Ref and underscore/non-underscore for roots/multiroots and min/multimin
This commit updates the use of Ptr vs Ref for input and output arguments both to various function wrappers and the ccall functions called within those wrappers for chapters 33-36 related to single dimensional and multidimensional root-finding and optimization.  This commit also begins to address various method redefinitions that occur due to the original underscore vs. non-underscore versions of files, wherein the underscored version of a file was auto-generated from the original Python script and the non-underscored version of a file contains hand-edited corrections.  This strategy will soon, in Julia v0.5, not be a viable approach due to the method redefinition it entails. In this commit, for chapters 33-36, the acutal use of underscored files has been eliminated as they are no longer being included.  The originally auto-generated __FILELIST.jl file has been hand-edited to make use of the non-underscored files for chapters 33-36, and any previously created non-underscored files for these chapters that were being included within the GSL.jl file are no longer included within that file.  With this package as it currently is designed, more and more of the autogenerated underscore files will be replaced with non-underscore files, and the underscored versions will no longer be included.  The are still present for reference, but just not used.  Longer term, these wrappers may be autogenerated using Clang.jl